requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.29.6
botocore==1.32.6
Django==4.2.7
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
Pillow==10.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.12.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
